pyproject.toml
[project]
name = "bpims"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=26.3.0",
    "django>=6.0.3",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "pip>=26.0.1",
    "psycopg2-binary>=2.9.11",
]
